Mount_Farmer/requirements.txt
Babel==2.6.0
blinker==1.4
certifi==2018.8.24
chardet==3.0.4
click==6.7
decorator==4.3.0
Flask==1.0.2
Flask-Babel==0.11.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.3.2
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.2
flup==1.0.3
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pbr==4.2.0
pip==6.1.1
pydocumentdb==2.3.2
python-wowapi==1.0
pytz==2018.5
requests==2.19.1
setuptools==40.2.0
six==1.11.0
speaklater==1.3
SQLAlchemy==1.2.11
sqlalchemy-migrate==0.11.0
sqlparse==0.2.4
Tempita==0.5.2
urllib3==1.23
Werkzeug==0.14.1
Whoosh==2.7.4
WTForms==2.2.1
